more-eks - separated the lb controller to it's own file, and shuffled things around
