+++ aws-foundation/modules/efs.py
@@ -1,9 +1,7 @@
 import pulumi
 import pulumi_aws as paws
+from pulumi_kubernetes.helm.v3 import Release, ReleaseArgs, RepositoryOptsArgs
 from config import AWSPulumiConfig
-import pprint
-
-pp = pprint.PrettyPrinter(indent=4)
 
 def _define_security_group(config: AWSPulumiConfig, vpc_data: dict) -> pulumi.Output:
     efs_sec = paws.ec2.SecurityGroup(f'{config.resource_prefix}-efs',
@@ -52,4 +50,28 @@ def define_efs(config: AWSPulumiConfig, vpc_data: dict) -> dict:
         'efs': efs
     }
 
-    
+def define_efs_controller(config: AWSPulumiConfig, k8s_provider: pulumi.Output, node_groups: list) -> Release:
+    repo_opt_args = RepositoryOptsArgs(
+        repo='https://kubernetes-sigs.github.io/aws-efs-csi-driver/'
+    )
+    release_args = ReleaseArgs(
+        chart='aws-efs-csi-driver',
+        create_namespace=True,
+        force_update=True,
+        namespace='kube-system',
+        repository_opts=repo_opt_args,
+        timeout=300,
+        version=config.efs.get('csi_driver').get('version')
+    )
+    release = Release(
+        resource_name=f'{config.resource_prefix}-efs-controller',
+        args=release_args,
+        opts=pulumi.ResourceOptions(
+            provider=k8s_provider, depends_on=node_groups
+        )
+    )
+
+    combined = pulumi.Output.all(release.name, release.version, release.status['status'])
+    pulumi.export(f'{config.resource_prefix}-efs-controller', combined.apply(lambda x: f'Name: {x[0]}, Version: {x[1]}, Status: {x[2]}'))
+
+    return release
